feat: adds agent mode for DOZZLE_MODE
